Reduzierung bei Dimensionssprung ohne T-Stück Fixes #37
