+++ metadata.txt
@@ -1,12 +1,9 @@
 [general]
 name=UTEC Massenermittlung (dev)
 
-version=0.6
+version=0.7
 changelog=
-    - Überarbeitung der Punktsuche und Darstellung
-    - Neue Spalte für Anmerkungen in Attributtabelle und Kartenhinweisen
-    - Sortierung der Attributtabelle verbessert (fragwürdige Punkte als erstes)
-    - Zusammenfassung der Ergebnisse bei hover-over Ergebnislayer
+    - Reduzierung bei Dimensionssprung ohne T-Stück
 
 description=Massenermittlung für Ferwärmenetze. Es werden Hausanschlüsse, Bögen, T-Stücke und Reduzierungen ausgegeben.
 qgisMinimumVersion=3.0
