+++ modules/t_intersection_analyzer.py
@@ -11,10 +11,10 @@ from qgis.core import (
 from qgis.PyQt.QtCore import QCoreApplication
 
 from . import constants as cont
-from .vector_analysis_tools import VectorAnalysisTools
+from .feature_creator import FeatureCreator
 
 
-class TIntersectionAnalyzer(VectorAnalysisTools):
+class TIntersectionAnalyzer(FeatureCreator):
     """A class to classify T-pieces and associated features."""
 
     def process_t_intersection(
@@ -77,7 +77,7 @@ class TIntersectionAnalyzer(VectorAnalysisTools):
         # fmt: off
         note_text: str = QCoreApplication.translate("feature_note", "Main pipe: {0}").format(" & ".join(note_parts))  # noqa: E501
         # fmt: on
-        created_count += self._create_t_piece(
+        created_count += self.create_t_piece(
             t_point, main_pipe_features, connecting_pipe, note_text
         )
 
@@ -193,57 +193,6 @@ class TIntersectionAnalyzer(VectorAnalysisTools):
         )
         return main_pipe, connecting_pipe
 
-    def _create_t_piece(
-        self,
-        point: QgsPointXY,
-        main_pipe: list[QgsFeature],
-        connecting_pipe: QgsFeature,
-        note: str,
-    ) -> int:
-        """Create a T-piece.
-
-        The T-piece dimension is determined by the largest dimension of the main
-        pipe and the dimension of the connecting pipe.
-
-        Args:
-            point: The geometry point for the T-piece.
-            main_pipe: A list of the two features forming the main pipe.
-            connecting_pipe: The feature for the connecting pipe.
-            note: A note to add to the feature's attributes.
-
-        Returns:
-            1 if the feature was created successfully, 0 otherwise.
-        """
-        all_features: list[QgsFeature] = [*main_pipe, connecting_pipe]
-        attrs: dict = {
-            cont.NewLayerFields.type.name: cont.Names.attr_val_type_t_piece,
-            cont.NewLayerFields.notes.name: note,
-        }
-        attrs |= self.get_connected_attributes(all_features)
-
-        if self.dim_field_name:
-            main_dims: list[int] = sorted(
-                [
-                    f.attribute(self.dim_field_name)
-                    for f in main_pipe
-                    if f.attribute(self.dim_field_name) is not None
-                ]
-            )
-            conn_dim: int | None = connecting_pipe.attribute(self.dim_field_name)
-
-            if main_dims and conn_dim is not None:
-                # T-piece uses the largest main pipe dim and the connecting pipe dim
-                t_dims_str: list[str] = [
-                    f"{cont.Names.dim_prefix}{main_dims[-1]}",
-                    f"{cont.Names.dim_prefix}{conn_dim}",
-                ]
-                dims_str: str = cont.Names.dim_separator.join(
-                    sorted(t_dims_str, reverse=True)
-                )
-                attrs[cont.NewLayerFields.dimensions.name] = dims_str
-
-        return 1 if self.create_feature(QgsGeometry.fromPointXY(point), attrs) else 0
-
     def _check_and_create_reducer(
         self,
         point: QgsPointXY,
@@ -274,7 +223,7 @@ class TIntersectionAnalyzer(VectorAnalysisTools):
             return 0
 
         if abs(dim_main_1 - dim_main_2) > 0:
-            return self._create_reducers(point, dim_main_1, dim_main_2, main_pipe)
+            return self.create_reducers(point, dim_main_1, dim_main_2, main_pipe)
 
         dim_conn: int | None = connecting_pipe.attribute(self.dim_field_name)
         if dim_conn is not None and dim_main_1 is not None and dim_conn > dim_main_1:
@@ -285,95 +234,3 @@ class TIntersectionAnalyzer(VectorAnalysisTools):
                 point, [*main_pipe, connecting_pipe], note=note_text
             )
         return 0
-
-    def _create_reducers(
-        self,
-        point: QgsPointXY,
-        dim_main_1: float,
-        dim_main_2: float,
-        main_pipe: list[QgsFeature],
-    ) -> int:
-        """Create reducers for a dimension change in the main pipe.
-
-        This method calculates how many reducers are needed and creates them
-        at appropriate distances from the intersection point.
-
-        Args:
-            point: The intersection point.
-            dim_main_1: Dimension of the first main pipe feature.
-            dim_main_2: Dimension of the second main pipe feature.
-            main_pipe: The two features forming the main pipe.
-
-        Returns:
-            The number of reducer features created.
-        """
-        large_dim: float = max(dim_main_1, dim_main_2)
-        small_dim: float = min(dim_main_1, dim_main_2)
-        smaller_dim_feature: QgsFeature = (
-            main_pipe[0] if dim_main_1 < dim_main_2 else main_pipe[1]
-        )
-        pref: str = cont.Names.dim_prefix
-
-        try:
-            large_idx: int = cont.PipeDimensions.diameters.index(large_dim)
-            small_idx: int = cont.PipeDimensions.diameters.index(small_dim)
-        except ValueError:
-            # fmt: off
-            note_text: str = QCoreApplication.translate("feature_note", "Non-standard pipe dimension detected for reducer.")  # noqa: E501
-            # fmt: on
-            return self.create_questionable_point(point, main_pipe, note=note_text)
-
-        dim_steps: int = large_idx - small_idx
-        if dim_steps <= 0:
-            return 0
-
-        num_reducers: int = (
-            dim_steps - 1
-        ) // cont.PipeDimensions.max_dim_jump_reducer + 1
-        created_count = 0
-
-        for i in range(num_reducers):
-            current_large_idx: int = (
-                large_idx - i * cont.PipeDimensions.max_dim_jump_reducer
-            )
-            current_small_idx: int = max(
-                small_idx,
-                large_idx - (i + 1) * cont.PipeDimensions.max_dim_jump_reducer,
-            )
-
-            dim_from: str = f"{pref}{cont.PipeDimensions.diameters[current_large_idx]}"
-            dim_to: str = f"{pref}{cont.PipeDimensions.diameters[current_small_idx]}"
-
-            distance: float = cont.Numbers.distance_t_reducer * (i + 1)
-            reducer_point: QgsPointXY | None = self._get_point_along_line(
-                point, smaller_dim_feature, distance
-            )
-            if not reducer_point:
-                continue
-
-            dims_str: str = cont.Names.dim_separator.join([dim_from, dim_to])
-            reducer_attrs = self.get_connected_attributes([smaller_dim_feature])
-            reducer_attrs.update(
-                {
-                    cont.NewLayerFields.type.name: cont.Names.attr_val_type_reducer,
-                    cont.NewLayerFields.dimensions.name: dims_str,
-                }
-            )
-            created_count += self.create_feature(
-                QgsGeometry.fromPointXY(reducer_point), reducer_attrs
-            )
-
-        return created_count
-
-    def _get_point_along_line(
-        self, start_point: QgsPointXY, feature: QgsFeature, distance: float
-    ) -> QgsPointXY | None:
-        """Get a point at a specific distance from the start point along a feature."""
-        other_endpoint: QgsPointXY | None = self.get_other_endpoint(
-            feature, start_point
-        )
-        if not other_endpoint:
-            return None
-
-        azimuth: float = start_point.azimuth(other_endpoint)
-        return start_point.project(distance, azimuth)
